KP-7131: Renamed numeric to Decimal

+++ src/Tests/WB.Tests.Unit.Designer/BoundedContexts/Designer/AddVariableHandlerTests/when_adding_variable_to_chapter.cs
@@ -71,7 +71,7 @@ namespace WB.Tests.Unit.Designer.BoundedContexts.Designer.AddStaticTextHandlerTe
         private static Guid responsibleId = Guid.Parse("DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD");
         private static string variableName = "name";
         private static string variableExpression = "expression";
-        private static VariableType variableType = VariableType.Long;
+        private static VariableType variableType = VariableType.Decimal;
         private static int index = 5;
     }
 }
